+++ http/getResponse.ts
@@ -4,18 +4,18 @@ import { GetProgressiveResponse } from '@cycler/task'
 
 function preprocessReqOptions(reqOptions: HTTPRequest): HTTPRequest & { method: string } {
   // TODO: make immutable
-  reqOptions.withCredentials = reqOptions.withCredentials || false;
-  reqOptions.redirects = typeof reqOptions.redirects === 'number' ? reqOptions.redirects : 5;
-  reqOptions.method = reqOptions.method || `get`;
+  reqOptions.withCredentials = reqOptions.withCredentials || false
+  reqOptions.redirects = typeof reqOptions.redirects === 'number' ? reqOptions.redirects : 5
+  reqOptions.method = reqOptions.method || `get`
   return reqOptions as HTTPRequest & { method: string }
 }
 type SuperAgentRequest = superagent.SuperAgentRequest
 function optionsToSuperagent(request: HTTPRequest): SuperAgentRequest {
-  const reqOptions = preprocessReqOptions(request);
+  const reqOptions = preprocessReqOptions(request)
 
-  const lowerCaseMethod = reqOptions.method.toLowerCase();
+  const lowerCaseMethod = reqOptions.method.toLowerCase()
 
-  let saRequest: SuperAgentRequest = (<any>superagent)[lowerCaseMethod](reqOptions.url);
+  let saRequest: SuperAgentRequest = (<any>superagent)[lowerCaseMethod](reqOptions.url)
 
   if (reqOptions.redirects) {
     saRequest.redirects(reqOptions.redirects)
@@ -23,6 +23,9 @@ function optionsToSuperagent(request: HTTPRequest): SuperAgentRequest {
   if (reqOptions.type) {
     saRequest.type(reqOptions.type)
   }
+  if (reqOptions.responseType) {
+    saRequest.responseType(reqOptions.responseType)
+  }  
   if (reqOptions.send) {
     saRequest.send(reqOptions.send)
   }
@@ -35,29 +38,29 @@ function optionsToSuperagent(request: HTTPRequest): SuperAgentRequest {
   if (reqOptions.withCredentials) {
     saRequest.withCredentials()
   }
-  if (typeof reqOptions.user === 'string' && typeof reqOptions.password === 'string') {
-    saRequest.auth(reqOptions.user, reqOptions.password);
+  if (reqOptions.auth) {
+    saRequest.auth(reqOptions.auth.user, reqOptions.auth.password)
   }
   if (reqOptions.headers) {
     for (let key in reqOptions.headers) {
       if (reqOptions.headers.hasOwnProperty(key)) {
-        saRequest.set(key, reqOptions.headers[key]);
+        saRequest.set(key, reqOptions.headers[key])
       }
     }
   }
   if (reqOptions.field) {
     for (let key in reqOptions.field) {
       if (reqOptions.field.hasOwnProperty(key)) {
-        saRequest.field(key, reqOptions.field[key]);
+        saRequest.field(key, reqOptions.field[key])
       }
     }
   }
   if (reqOptions.attach) {
     reqOptions.attach.forEach(a => {
-      saRequest.attach(a.name, a.path, a.filename);
+      saRequest.attach(a.name, a.path, a.filename)
     })
   }
-  return saRequest;
+  return saRequest
 }
 
 export function normalizeRequest(request: HTTPRequestInput): HTTPRequest {
@@ -75,23 +78,23 @@ export const getProgressiveResponse:
   GetProgressiveResponse<HTTPRequest, HTTPResponse, any> =
   (request, observer, onDispose) => {
     try {
-      let saRequest = optionsToSuperagent(request);
+      let saRequest = optionsToSuperagent(request)
       if (request.progress) {
         saRequest.on('progress', observer.next)
       }
 
       saRequest.end((err: any, res: HTTPResponse) => {        
         if (err) {
-          observer.error(err);
+          observer.error(err)
         } else {
-          observer.next(res);
-          observer.complete();
+          observer.next(res)
+          observer.complete()
         }
-      });
+      })
       onDispose(() => {
         saRequest.abort()
       })
     } catch (err) {
-      observer.error(err);
+      observer.error(err)
     }
   }