+++ http/CHANGELOG.md
@@ -3,5 +3,7 @@
 1.0.0.pre.0
 
 - @cycler/task `InputTaskSource` support
+- `responseType` option
+- `auth` option
 
 1.0.0-unified.0 (2017-03-21)